Refactor: tidy type declaration files
