+++ package.d.ts
@@ -1,0 +1,1 @@
+export const PackageRoot: string;
