--- src/types/package-root.d.ts
@@ -1,4 +1,0 @@
-module "#package-root" {
-  export const PackageRoot: string;
-}
-//# sourceMappingURL=package-root.d.ts.map
